requirements.txt
python-dotenv == 0.21.1
fastapi == 0.89.1
uvicorn == 0.20.0
pydantic == 1.10.4
sqlmodel == 0.0.8
psycopg2-binary == 2.9.5
asyncpg == 0.27.0
python-dateutil == 2.8.2